backend/requirements.txt
fastapi
uvicorn[standard]
supabase
python-dotenv
pydantic
python-multipart
python-jose[cryptography]
passlib[bcrypt]
ibm-watsonx-ai
